refactor(dify-ui): rename radio family entrypoint
