+++ packages/dify-ui/package.json
@@ -109,9 +109,9 @@
       "types": "./src/pagination/index.tsx",
       "import": "./src/pagination/index.tsx"
     },
-    "./radio": {
-      "types": "./src/radio/index.tsx",
-      "import": "./src/radio/index.tsx"
+    "./radio-group": {
+      "types": "./src/radio-group/index.tsx",
+      "import": "./src/radio-group/index.tsx"
     },
     "./popover": {
       "types": "./src/popover/index.tsx",
