+++ web/app/components/app/app-access-control/__tests__/access-control-item.spec.tsx
@@ -1,4 +1,4 @@
-import { RadioGroup } from '@langgenius/dify-ui/radio'
+import { RadioGroup } from '@langgenius/dify-ui/radio-group'
 import { render, screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import { useState } from 'react'
