+++ web/app/components/explore/category.tsx
@@ -1,7 +1,7 @@
 'use client'
 import type { AppCategory } from '@/models/explore'
 import { cn } from '@langgenius/dify-ui/cn'
-import { RadioGroup, RadioItem } from '@langgenius/dify-ui/radio'
+import { RadioGroup, RadioItem } from '@langgenius/dify-ui/radio-group'
 import { useTranslation } from 'react-i18next'
 import exploreI18n from '@/i18n/en-US/explore.json'
