+++ web/app/components/integrations/permission-quick-panel.tsx
@@ -2,7 +2,7 @@
 
 import type { Permissions } from '@/app/components/plugins/types'
 import { cn } from '@langgenius/dify-ui/cn'
-import { RadioGroup, RadioItem } from '@langgenius/dify-ui/radio'
+import { RadioGroup, RadioItem } from '@langgenius/dify-ui/radio-group'
 import { useTranslation } from 'react-i18next'
 import { PluginSidecarPanel } from '@/app/components/plugins/plugin-page/plugin-sidecar-panel'
 import { PermissionType } from '@/app/components/plugins/types'
