+++ web/app/components/tools/edit-custom-collection-modal/config-credentials.tsx
@@ -15,7 +15,7 @@ import {
 import { Field, FieldItem, FieldLabel } from '@langgenius/dify-ui/field'
 import { Fieldset, FieldsetLegend } from '@langgenius/dify-ui/fieldset'
 import { Input } from '@langgenius/dify-ui/input'
-import { Radio, RadioGroup } from '@langgenius/dify-ui/radio'
+import { Radio, RadioGroup } from '@langgenius/dify-ui/radio-group'
 import {
   ScrollArea,
   ScrollAreaContent,
