+++ web/features/new-rag/components/create-knowledge-dialog-parts.tsx
@@ -3,7 +3,7 @@
 import type { ReactNode } from 'react'
 import type { NewKnowledgeStartMode } from '../routes'
 import { cn } from '@langgenius/dify-ui/cn'
-import { RadioControl, RadioItem } from '@langgenius/dify-ui/radio'
+import { RadioControl, RadioItem } from '@langgenius/dify-ui/radio-group'
 import { useId } from 'react'
 import { useTranslation } from 'react-i18next'
